backend/requirements.txt
fastapi
uvicorn[standard]
websockets
openai
pydantic
python-multipart
python-dotenv
pinecone
pandas
numpy
google-generativeai
